+++ core/commands.lua
@@ -11,6 +11,7 @@ local tinsert, tsort = table.insert, table.sort
 
 -- string formats
 local STRING_COLOR = "|cffb3ff19%s|r"   --"|cff00ff96%s|r"
+local AddOnCommands = {}
 
 local function CmdSplit(cmd)
     if (cmd:find("%s")) then
@@ -88,10 +89,6 @@ SlashCmdList["FILGER"] = function(cmd)
             frame.Backdrop:Hide()
             frame.text:Hide()
         end
-    elseif (arg1 == "reset") and (Config["Plugins"].Tracer) then
-        Filger:ResetZone()
-    elseif (arg1 == "tracer") and (Config["Plugins"].Tracer) then
-        Filger:TracerEnable()
     elseif (arg1 == "") or (arg1 == "help") then
         print(" ")
         print(STRING_COLOR:format("Filger Commands:"))
@@ -101,5 +98,9 @@ SlashCmdList["FILGER"] = function(cmd)
         print(STRING_COLOR:format("reset"), "Reset saved variables.")
         print(STRING_COLOR:format("tracer"), "Start/Stop filger aura tracer.")
         print(" ")
+    elseif (AddOnCommands[arg1]) then
+        AddOnCommands[arg1](arg2)
     end
 end
+
+Filger.AddOnCommands = AddOnCommands