+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@
 + Added: WoLK dungeons and raid tables.
 + Added: update BoA dungeon and raid tables.
 + Added: Tracer plugins (help to build spell tables by saving buffs/debuffs that affect player)
++ Changed: Tracer plugin enhanced (added automatic option).
 
 ## Changes in 2.3-release